# api-rest-cine/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api rest en GO de un cine",
        "title": "Cine",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/",
    "paths": {
...
        "/api/v1/horario": {
            "get": {
                "description": "Hace la petición para obtener un horario mediante la salaID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtiene un item por PeliculaID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sala a buscar",
                        "name": "pelicula_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/horario/{id}": {
            "get": {
                "description": "Hace la petición para obtener un horario por su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtiene un horario en especifico",
                "operationId": "Buscar un horario por su id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id del horario a buscar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/precio": {
            "get": {
                "description": "Hace la petición para obtener los precios mediante la salaID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Precios"
                ],
                "summary": "Obtiene precios con la llave foranea SalaID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "llave forenea de sala en precios",
                        "name": "sala_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/precio/{id}": {
            "get": {
                "description": "Hace la petición para obtener una solo precio por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Precios"
                ],
                "summary": "Obtiene un precio en especifico",
                "operationId": "Buscar precio por su ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del precio a buscar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/promocion/{id}": {
            "get": {
                "description": "Hace la petición para obtener una promocion por su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promociones"
                ],
                "summary": "Obtiene una promocion en especifico",
                "operationId": "Buscar una promocion por su id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la promocion a buscar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sala/{ID}": {
            "get": {
                "description": "Hace la petición para obtener una sola por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salas"
                ],
                "summary": "Obtiene una sala en especifico",
                "operationId": "Buscar sala por su ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la sala a buscar",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/createpelicula": {
            "post": {
                "description": "Insertar un nueva película en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peliculas"
                ],
                "summary": "Insertar una película en la base de datos",
                "responses": {
                    "201": {
                        "description": "Éxito",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pelicula": {
            "get": {
                "description": "Hace la petición para obtener la lista de películas que estén en cartelera",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peliculas"
                ],
                "summary": "Obtiene una lista de películas en cartelera",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "columna a comprobar",
                        "name": "estreno",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/peliculas/{id}": {
            "get": {
                "description": "Hace la petición para obtener una sola pelicula por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peliculas"
                ],
                "summary": "Obtiene peliculas",
                "operationId": "Buscar elemento por su ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la película a buscar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}